requirements.txt
Django==2.2.5
django-crispy-forms==1.7.2
django-tempus-dominus==5.1.2.8
pytz==2019.2
sqlparse==0.3.0
